Rename with proper capitalization
